Move ScientificPublications
 - Создал новое приложение "scientific_publications"
 - Переместил всё, что связано с backend стороной у научных публикаций в новую папку приложения department/scientific_publications из папки editable-pages.
- Были обновлены URL для API-endpoints
- Сделаны необходимые миграции. Удаление их из прошлой папки и создание в новой папке приложения.
- Были поправлены миграции news.
