+++ backend/db.local.env
@@ -1,5 +1,5 @@
 POSTGRES_USER=postgres
-POSTGRES_PASSWORD=1234
+POSTGRES_PASSWORD=
 POSTGRES_DB=postgres
 TG_BOT_TOKEN=
 TG_WEBHOOK_HOST=
