+++ backend/department/scientific_publications/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ScientificPublicationsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'department.scientific_publications'
+    verbose_name = 'Научные публикации'
