+++ backend/editable_pages/serializers.py
@@ -1,21 +1,9 @@
 from rest_framework import serializers
 
-from editable_pages.models import EditablePage, ScientificPublication, Tag
+from editable_pages.models import EditablePage
 
 
 class EditablePageSerializer(serializers.ModelSerializer):
     class Meta:
         model = EditablePage
-        fields = '__all__'
-
-class TagSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Tag
-        fields = '__all__'
-
-class ScientificPublicationSerializer(serializers.ModelSerializer):
-    tags = TagSerializer(many=True)
-
-    class Meta:
-        model = ScientificPublication
         fields = '__all__'
